Accept alternate annotation for using external IP
In addition to the "external-dns.alpha.kubernetes.io/endpoints-type"
annotation honored for headless Services, accept the similar
".../use-external-host-ip" annotation to specify whether ExternalDNS
should create DNS records pointing at pods' IP addresses or—when the
"publish-host-ip" feature is active and this annotation's value is
"true"—the external IP addresses of the nodes hosting the pods
selected by the Service.

If the annotation is present but bears any value other than "true,"
ExternalDNS treats it as if the annotation was absent, using either
the hosting nodes' first (and usually internal) host IP address when
the "publish-host-ip" feature is active or each pod's IP address
otherwise.