Skip creating some records with mixed identifiers
If ExternalDNS considers creating a DNS record with an empty set
identifier, but there are already existing records with the same DNS
name but nonempty set identifiers, skip creation of the record with
the empty set identifier.

Doing so allows ExternalDNS to preserve DNS records with set
identifiers managed by other systems that still must be mentioned in
sources like Kubernetes Ingress objects, even if those DNS names sit
within zones in which we expect ExternalDNS to operate.